coprocessor/* : split dag and old select into independent mod (#2128)
