+++ src/coprocessor/codec/datum.rs
@@ -24,7 +24,7 @@ use util::escape;
 use util::codec::{number, bytes};
 use util::codec::number::NumberDecoder;
 use util::codec::bytes::BytesEncoder;
-use super::super::xeval::EvalContext;
+use coprocessor::select::xeval::EvalContext;
 use super::{Result, convert};
 use super::mysql::{self, Duration, DEFAULT_FSP, MAX_FSP, Decimal, DecimalEncoder, DecimalDecoder,
                    Time, Json, JsonEncoder, JsonDecoder, PathExpression, parse_json_path_expr};
@@ -1328,7 +1328,7 @@ mod test {
             (Datum::Dec(0u64.into()), Some(false)),
         ];
         use chrono::FixedOffset;
-        use coprocessor::xeval::EvalContext;
+        use coprocessor::select::xeval::EvalContext;
 
         let ctx = EvalContext {
             tz: FixedOffset::east(0),
