+++ src/coprocessor/dag/dag.rs
@@ -16,26 +16,25 @@ use std::time::Instant;
 
 use tipb::executor::{Executor, ExecType};
 use tipb::schema::ColumnInfo;
-use tipb::select::{DAGRequest, Chunk};
-use kvproto::coprocessor::KeyRange;
+use tipb::select::{DAGRequest, SelectResponse, RowMeta};
+use kvproto::coprocessor::{Response, KeyRange};
 use kvproto::kvrpcpb::IsolationLevel;
+use protobuf::{Message as PbMsg, RepeatedField};
 
+use coprocessor::codec::mysql;
+use coprocessor::codec::datum::{DatumEncoder, Datum};
+use coprocessor::select::xeval::EvalContext;
+use coprocessor::{Result, Error};
+use coprocessor::endpoint::{check_if_outdated, to_pb_error, get_chunk, get_pk, REQ_TYPE_DAG};
 use storage::{Snapshot, SnapshotStore, Statistics};
-use super::xeval::EvalContext;
-use super::{Result, Error};
-use super::executor::Executor as DAGExecutor;
-use super::executor::table_scan::TableScanExecutor;
-use super::executor::index_scan::IndexScanExecutor;
-use super::executor::selection::SelectionExecutor;
-use super::executor::aggregation::AggregationExecutor;
-use super::executor::topn::TopNExecutor;
-use super::executor::limit::LimitExecutor;
+
+use super::executor::{Executor as DAGExecutor, TableScanExecutor, IndexScanExecutor,
+                      SelectionExecutor, AggregationExecutor, TopNExecutor, LimitExecutor, Row};
 
 pub struct DAGContext<'s> {
-    pub deadline: Instant,
-    pub columns: Rc<Vec<ColumnInfo>>,
-    pub has_aggr: bool,
-    pub chunks: Vec<Chunk>,
+    deadline: Instant,
+    columns: Rc<Vec<ColumnInfo>>,
+    has_aggr: bool,
     req: DAGRequest,
     ranges: Vec<KeyRange>,
     snap: &'s Snapshot,
@@ -59,16 +58,57 @@ impl<'s> DAGContext<'s> {
             snap: snap,
             has_aggr: false,
             eval_ctx: eval_ctx,
-            chunks: vec![],
             isolation_level: isolation_level,
         }
     }
 
-    pub fn get_output_offsets(&self) -> &[u32] {
-        self.req.get_output_offsets()
+    pub fn handle_request(mut self, statistics: &'s mut Statistics) -> Result<Response> {
+        try!(self.validate_dag());
+        let mut exec = try!(self.build_dag(statistics));
+        let mut chunks = vec![];
+        loop {
+            match exec.next() {
+                Ok(Some(row)) => {
+                    try!(check_if_outdated(self.deadline, REQ_TYPE_DAG));
+                    let mut chunk = get_chunk(&mut chunks);
+                    let length = chunk.get_rows_data().len();
+                    if self.has_aggr {
+                        chunk.mut_rows_data().extend_from_slice(&row.data.value);
+                    } else {
+                        let value =
+                            try!(inflate_cols(&row, &self.columns, self.req.get_output_offsets()));
+                        chunk.mut_rows_data().extend_from_slice(&value);
+                    }
+                    let mut meta = RowMeta::new();
+                    meta.set_handle(row.handle);
+                    meta.set_length((chunk.get_rows_data().len() - length) as i64);
+                    chunk.mut_rows_meta().push(meta);
+                }
+                Ok(None) => {
+                    let mut resp = Response::new();
+                    let mut sel_resp = SelectResponse::new();
+                    sel_resp.set_chunks(RepeatedField::from_vec(chunks));
+                    let data = box_try!(sel_resp.write_to_bytes());
+                    resp.set_data(data);
+                    return Ok(resp);
+                }
+                Err(e) => {
+                    if let Error::Other(_) = e {
+                        let mut resp = Response::new();
+                        let mut sel_resp = SelectResponse::new();
+                        sel_resp.set_error(to_pb_error(&e));
+                        resp.set_data(box_try!(sel_resp.write_to_bytes()));
+                        resp.set_other_error(format!("{}", e));
+                        return Ok(resp);
+                    } else {
+                        return Err(e);
+                    }
+                }
+            }
+        }
     }
 
-    pub fn validate_dag(&mut self) -> Result<()> {
+    fn validate_dag(&mut self) -> Result<()> {
         let execs = self.req.get_executors();
         let first = try!(execs.first()
             .ok_or_else(|| Error::Other(box_err!("has no executor"))));
@@ -94,10 +134,10 @@ impl<'s> DAGContext<'s> {
 
     // seperate first exec build action from `build_dag`
     // since it will generte mutable conflict when putting together
-    pub fn build_first(&'s self,
-                       mut first: Executor,
-                       statistics: &'s mut Statistics)
-                       -> Box<DAGExecutor + 's> {
+    fn build_first(&'s self,
+                   mut first: Executor,
+                   statistics: &'s mut Statistics)
+                   -> Box<DAGExecutor + 's> {
         let store = SnapshotStore::new(self.snap, self.req.get_start_ts(), self.isolation_level);
 
         match first.get_tp() {
@@ -117,7 +157,7 @@ impl<'s> DAGContext<'s> {
         }
     }
 
-    pub fn build_dag(&'s self, statistics: &'s mut Statistics) -> Result<Box<DAGExecutor + 's>> {
+    fn build_dag(&'s self, statistics: &'s mut Statistics) -> Result<Box<DAGExecutor + 's>> {
         let mut execs = self.req.get_executors().to_vec().into_iter();
         let mut src = self.build_first(execs.next().unwrap(), statistics);
         for mut exec in execs {
@@ -149,4 +189,32 @@ impl<'s> DAGContext<'s> {
         }
         Ok(src)
     }
+}
+
+#[inline]
+fn inflate_cols(row: &Row, cols: &[ColumnInfo], output_offsets: &[u32]) -> Result<Vec<u8>> {
+    let data = &row.data;
+    // TODO capacity is not enough
+    let mut values = Vec::with_capacity(data.value.len());
+    for offset in output_offsets {
+        let col = &cols[*offset as usize];
+        let col_id = col.get_column_id();
+        match data.get(col_id) {
+            Some(value) => values.extend_from_slice(value),
+            None if col.get_pk_handle() => {
+                let pk = get_pk(col, row.handle);
+                box_try!(values.encode(&[pk], false));
+            }
+            None if col.has_default_val() => {
+                values.extend_from_slice(col.get_default_val());
+            }
+            None if mysql::has_not_null_flag(col.get_flag() as u64) => {
+                return Err(box_err!("column {} of {} is missing", col_id, row.handle));
+            }
+            None => {
+                box_try!(values.encode(&[Datum::Null], false));
+            }
+        }
+    }
+    Ok(values)
 }
