+++ src/coprocessor/dag/executor/aggregation.rs
@@ -18,14 +18,15 @@ use tipb::executor::Aggregation;
 use tipb::expression::Expr;
 use util::collections::{HashMap, HashMapEntry as Entry};
 
+use coprocessor::codec::table::RowColsDict;
+use coprocessor::codec::datum::{self, Datum, DatumEncoder, approximate_size};
+use coprocessor::endpoint::SINGLE_GROUP;
+use coprocessor::select::aggregate::{self, AggrFunc};
+use coprocessor::select::xeval::{Evaluator, EvalContext};
+use coprocessor::metrics::*;
+use coprocessor::Result;
+
 use super::{Executor, Row, ExprColumnRefVisitor, inflate_with_col_for_dag};
-use super::super::codec::table::RowColsDict;
-use super::super::codec::datum::{self, Datum, DatumEncoder, approximate_size};
-use super::super::xeval::{Evaluator, EvalContext};
-use super::super::endpoint::SINGLE_GROUP;
-use super::super::aggregate::{self, AggrFunc};
-use super::super::metrics::*;
-use super::super::Result;
 
 pub struct AggregationExecutor<'a> {
     group_by: Vec<Expr>,
